+++ decidim-elections/app/permissions/decidim/elections/admin/permissions.rb
@@ -19,10 +19,12 @@ module Decidim
             case permission_action.action
             when :create, :read
               allow!
-            when :update
+            when :update, :publish
               toggle_allow(election)
             when :delete
               allow_if_not_started
+            when :unpublish
+              allow_if_not_started_or_finished
             end
           end
 
@@ -37,6 +39,10 @@ module Decidim
 
         def allow_if_not_started
           toggle_allow(election && !election.started?)
+        end
+
+        def allow_if_not_started_or_finished
+          toggle_allow(election && (!election.started? || election.finished?))
         end
       end
     end